Reduced from bigint to int in locationDbSchema.sql
Former-commit-id: 50863678a0200e1ebe34d5923f1bd4d43752cb11

+++ config/locationDbSchema.sql
@@ -26,7 +26,7 @@
 DROP TABLE IF EXISTS `location_alternative_names`;
 
 CREATE TABLE `location_alternative_names` (
-  `locationId` bigint(20) unsigned NOT NULL COMMENT 'The id of the location.',
+  `locationId` int(11) unsigned NOT NULL COMMENT 'The id of the location.',
   `alternativeName` varchar(200) DEFAULT NULL COMMENT 'An alternative name used for the location.',
   `language` char(2) DEFAULT NULL COMMENT 'The language for this alternative name, in ISO 639-1 format. NULL means no specified language.',
   UNIQUE KEY `idNameLangUnique` (`locationId`,`alternativeName`,`language`),
@@ -43,7 +43,7 @@ CREATE TABLE `location_alternative_names` (
 DROP TABLE IF EXISTS `locations`;
 
 CREATE TABLE `locations` (
-  `id` bigint(20) unsigned NOT NULL DEFAULT '0' COMMENT 'The id of the location.',
+  `id` int(11) unsigned NOT NULL DEFAULT '0' COMMENT 'The id of the location.',
   `type` varchar(20) NOT NULL DEFAULT '' COMMENT 'The type of the location.',
   `name` varchar(255) NOT NULL COMMENT 'The primary name of the location.',
   `latitude` double(8,5) DEFAULT NULL COMMENT 'The latitude of the location.',
